desktop-app/requirements.txt
pywebview>=5.0
yt-dlp>=2024.01.01
pyinstaller>=6.0
